+++ CHANGELOG.md
@@ -18,7 +18,8 @@
 
 #### Builtin
 
-- **Breaking Change**: change help message format of builtin commands
+- **Breaking Change**: change a help message format of builtin commands
+- builtin help command also shows user-defined commands
 - improve ``String#split`` method
     - ``delim`` param is now optional (pass empty string by default)
     - add ``limit`` param